# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.4.0
  description: PerformanceIQ Football Performance SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams with nested subteams
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamWithSubteams"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    patch:
      operationId: updateTeam
      tags: [teams]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    delete:
      operationId: deleteTeam
      tags: [teams]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
  /teams/{id}/subteams:
    get:
      operationId: listTeamSubteams
      tags: [teams]
      summary: List subteams for a team
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: List of subteams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subteam"
    post:
      operationId: createSubteam
      tags: [teams]
      summary: Create a subteam
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubteamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subteam"
        "404":
          description: Team not found
  /subteams/{id}:
    delete:
      operationId: deleteSubteam
      tags: [teams]
      summary: Delete a subteam
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    Team:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        clubName: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        season: { type: ["string", "null"] }
        parentTeamId: { type: ["integer", "null"] }
        createdAt: { type: string }
    TeamWithSubteams:
      type: object
      required: [id, name, createdAt, subteams]
      properties:
        id: { type: integer }
        name: { type: string }
        clubName: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        season: { type: ["string", "null"] }
        parentTeamId: { type: ["integer", "null"] }
        createdAt: { type: string }
        subteams:
          type: array
          items:
            $ref: "#/components/schemas/Subteam"
    TeamInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        clubName: { type: string }
        category: { type: string }
        season: { type: string }
        parentTeamId: { type: integer }
    Subteam:
      type: object
      required: [id, teamId, name, createdAt]
      properties:
        id: { type: integer }
        teamId: { type: integer }
        name: { type: string }
        createdAt: { type: string }
    SubteamInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
...