# lib/api-zod/src/generated/types/subteam.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PerformanceIQ Football Performance SaaS API
 * OpenAPI spec version: 0.4.0
 */

export interface Subteam {
  id: number;
  teamId: number;
  name: string;
  createdAt: string;
}